requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.4.0
librosa==0.10.1
python-multipart==0.0.9
sqlalchemy==2.0.36
alembic==1.14.0
psycopg[binary]==3.2.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
python-dotenv==1.0.1
